Enable partial reading of simple files
The SimpleFile module kept track of the current position in the
file (so that it could resume reading rather than reading from
the start each time), but this facility was disabled with the
"_filepos_last_read" method always returning 0.

However we need to protect against reading partial lines (potentially
not beginning at the start).  Therefore only accept lines which
contain the line feed character.  The file position is only tracked after
such lines are read.